+++ src/fairseq2/recipes/lm/_online_finetune/__init__.py
@@ -4,4 +4,6 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from fairseq2.recipes.lm._online_finetune._remote_model import RayActorConfig, VllmEngineArgs
+from fairseq2.recipes.lm._online_finetune._remote_model import (
+    RayActorConfig as RayActorConfig,
+)
